test sum int

+++ dimabe_export_order/models/stock_picking.py
@@ -256,7 +256,7 @@ class StockPicking(models.Model):
         for item in result:
             if item.name == self.origin:
                 for i in item.order_line:
-                    models._logger.error(sum(i.price_unit * i.product_uom_qty))
+                    models._logger.error(sum(int(i.price_unit * i.product_uom_qty)))
                     self.total_value = (i.price_unit * i.product_uom_qty) + self.freight_value + self.safe_value
 
     @api.model
